update al slow contorl infrastructure to use microseconds scale into the scheduler step delay computation

+++ chaos/common/chaos_types.h
@@ -123,8 +123,8 @@ namespace chaos {
 						//! If true the user modification to the features are denied
 						//bool lockedOnUserModification;
 
-						//! Command specific delay beetween two sequence of the commadn step (acquire->correlation)
-						uint32_t featureSchedulerStepsDelay;
+						//! Command specific delay beetween two sequence of the commadn step (acquire->correlation) in microseocnds
+						uint64_t featureSchedulerStepsDelay;
 						
 						//! Checker time
 						/*!
